requirements.txt
click==8.1.3
colorama==0.4.6
cycler==0.11.0
Flask==2.2.5
fonttools==4.38.0        
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.21.6
packaging==23.1
Pillow==9.5.0
pyngrok==6.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
typing_extensions==4.6.1
Werkzeug==2.2.3
wordcloud==1.9.2
zipp==3.15.0
